backend: ref- delete & edit usersPost

+++ backend/src/controllers/post.controller.js
@@ -59,11 +59,19 @@ const getPost = async (req, res) => {
 
 const deletePost = async (req, res) => {
   const postId = req.params.postId;
+  const userId = req.userId;
   try {
     const post = await Post.findById(postId);
 
     if (!post) {
       throw new Error("Post not found");
+    }
+    const user = await User.findById(userId).select("-password");
+    const usersPost = user.posts.find((blog) => blog._id == postId);
+    if (!usersPost) {
+      throw new Error(
+        "You are not the author. You cannot delete or edit this post."
+      );
     }
 
     await post.deleteOne();
@@ -76,7 +84,10 @@ const deletePost = async (req, res) => {
 
 const editPost = async (req, res) => {
   const postId = req.params.postId;
-  const { title, category, body, image, readingTime } = req.body;
+  const userId = req.userId;
+
+  const { title, category, body, readingTime } = req.body;
+  const image = req.file;
 
   if (!title || !category || !body || !image || !readingTime) {
     throw new Error("All fields are required");
@@ -86,16 +97,24 @@ const editPost = async (req, res) => {
   if (!post) {
     throw new Error("Post not found");
   }
+
+  const user = await User.findById(userId).select("-password");
+  const usersPost = user.posts.find((blog) => blog._id == postId);
+  if (!usersPost) {
+    throw new Error(
+      "You are not the author. You cannot delete or edit this post."
+    );
+  }
+
   try {
     post.title = title;
     post.category = category;
     post.body = body;
-    post.image = image;
+    post.image = `images/${image.filename}`;
     post.readingTime = readingTime;
     post.lastUpdate = Date.now();
 
     await post.save();
-
     res.status(200).json({ success: true, post });
   } catch (error) {
     res.status(500).json({ success: false, message: error.message });
